Feat: 개발환경에 따른 코드 변경, 이미지 표출 방법 변경
- local, prod 환경에 따라 사용하는 Service가 다르도록 interface와 implements로 분리
- 이미지 표출 방법도 환경에 따라 다르게 설정
